refactor(forgesvc): move board visibility/share seam to projectsync_share.go (#963)
Pure same-package code motion, move-only (PRD #963 M1). Moves the five
GitHub Projects v2 board-access methods — GetVisibility, RepoOwnerType,
SetVisibility, ShareWithUser, Unshare — verbatim (bodies and doc comments
byte-identical) out of projectsync.go into a new sibling file
projectsync_share.go, same package forgesvc. No exported-API change, no
behavior change, no test edit; imports settled by the compiler. Verified
byte-identical motion (121 moved lines diff-clean) and go build/vet green.
